fix auth

+++ ASP.NET-API Configuration/DataAccess/DAOs/AccessTokenDAO.cs
@@ -1,5 +1,6 @@
 ﻿using BusinessObject;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.ChangeTracking;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -28,22 +29,34 @@ namespace DataAccess.DAOs
 			}
 		}
 
-		internal async Task AddAccessTokenAsync(AccessToken accessToken)
+		internal async Task<AccessToken> AddAccessTokenAsync(AccessToken accessToken)
 		{
 			using (ApiContext context = new ApiContext())
 			{
-				 await context.AccessToken.AddAsync(accessToken);
-				context.SaveChanges();
+				EntityEntry<AccessToken> token = await context.AccessToken.AddAsync(accessToken);
+				await context.SaveChangesAsync();
+				return token.Entity;
 			}
 		}
 
-		internal async Task RemoveAllAccessTokenByUserIdAsync(string? userId)
+		internal async Task RemoveAllAccessTokenUserAsync(string? userId,  string? jwtId)
 		{
 			using (ApiContext context = new ApiContext())
 			{
-				var accessTokens = await context.AccessToken.Where(x => x.UserId == int.Parse(userId ?? "0")).ToListAsync();
-				context.AccessToken.RemoveRange(accessTokens);
-				context.SaveChanges();	
+				List<AccessToken> accessTokens = new List<AccessToken>();
+				if(!string.IsNullOrWhiteSpace(userId)) 
+				{
+					 accessTokens = await context.AccessToken.Where(x => x.UserId == int.Parse(userId)).ToListAsync();
+				}
+				else
+				{
+					accessTokens = await context.AccessToken.Where(x => x.JwtId == jwtId).ToListAsync();
+				}
+				if(accessTokens.Count > 0)
+				{
+					context.AccessToken.RemoveRange(accessTokens);
+					await context.SaveChangesAsync();
+				}
 			}
 		}
